% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minba.R
\name{minba()}
\alias{minba()}
\alias{minba}
\title{Determining the Minimal Background Area for Species Distribution Models}
\usage{
minba(occ = NULL, varbles = NULL, wd = NULL, prj = NULL,
  num_bands = 10, n_rep = 3, BI_part = NULL, BI_tot = NULL,
  SD_BI_part = NULL, SD_BI_tot = NULL)
}
\arguments{
\item{occ}{Data set with presences (occurrences). A data frame with 3 columns: long, lat and species name (in this order)}

\item{varbles}{An raster brick of the independent variables, or a directory where the rasters are. It will use all the rasters in the folder. Supported: .tif and .hdr Labelled .bil}

\item{wd}{A directory to save the results}

\item{prj}{Coordinates system (e.g. "4326" is WGS84; check \url{http://spatialreference.org/} )}

\item{num_bands}{Number of buffers}

\item{n_rep}{Number of replicates}

\item{BI_part}{Maximum Boyce Index Partial to stop the process if reached}

\item{BI_tot}{Maximum Boyce Index Total to stop the process if reached}

\item{SD_BI_part}{Minimum SD of the Boyce Index Partial to stop the process if reached (last 3 buffers)}

\item{SD_BI_tot}{Minimum SD of the Boyce Index Total to stop the process if reached (last 3 buffers)}
}
\value{
\code{selfinfo_mod_}, \code{info_mod_} and \code{info_mod_means_} (all followed by the name of the species). The first two tables are merely informative about how the modelling process has been developed and the results of each model. Whereas \code{info_mod_means_} shows the means of the n models run for each buffer
}
\description{
A versatile tool that aims at (1) defining what is the minimum or optimal background extent necessary to fit good partial species distribution models and/or (2) determining if the background area used to fit a partial species distribution model is reliable enough to extract ecologically relevant conclusions from it. See Rotllan-Puig, X. & Traveset, A. (2019)
}
\details{
Please check the article 'Determining the Minimal Background Area for Species Distribution Models: MinBAR Package' for further details on how to use this package, examples, etc.
}
\examples{

# Saving maxent.jar
if(!file.exists(paste0(tempdir(), "/dismo/java/maxent.jar"))){
dir.create(paste0(tempdir(), "/dismo/"))
dir.create(paste0(tempdir(), "/dismo/java"))
file.copy(system.file("maxent.jar", package="MinBAR"), paste0(tempdir(), "/dismo/java/maxent.jar"))
}
MinBAR:::minba(occ = sprecords, varbles = bioscrop, wd = tempdir(), prj = 4326, num_bands = 3)


}
\references{
Rotllan-Puig, X. & Traveset, A. 2019. Determining the Minimal Background Area for Species Distribution Models: MinBAR Package. bioRxiv. 571182. DOI: 10.1101/571182
}
\author{
Xavier Rotllan-Puig & Anna Traveset
}
