% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sprecords}
\alias{sprecords}
\title{Presences (occurrences) of Linaria alpina}
\format{A data frame with 1064 rows and 3 variables.
\describe{
  \item{decimalLongitude}{DecimalLongitude, in degrees}
  \item{decimalLatitude}{DecimalLatitude, in degrees}
  \item{species}{Name of the species}
}}
\source{
\url{https://www.gbif.org/}
}
\usage{
sprecords
}
\description{
A dataset containing the presences (1064) of Linaria alpina in Europe and North Africa.
Coord. ref. : +init=EPSG:4326 +proj=longlat +datum=WGS84 +no_defs +ellps=WGS84 +towgs84=0,0,0.
}
\references{
GBIF.org (07 March 2018) GBIF Occurrence Download https://doi.org/10.15468/dl.phqgk3.
}
\keyword{datasets}
