% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minba.R
\name{minba()}
\alias{minba()}
\alias{minba}
\title{Downloading species ocurrence data (presences) from Bioatles}
\usage{
minba(occ = NULL, varbles = NULL, prj = NULL, num_bands = 10,
  n_times = 3, BI_part = NULL, BI_tot = NULL, SD_BI_part = NULL,
  SD_BI_tot = NULL)
}
\arguments{
\item{occ}{Data set with presences (occurrences). A csv file with 3 columns: long, lat and species name (in this order)}

\item{varbles}{A directory where the independent variables (rasters) are. It will use all of them in the folder. Supported: .tif and .hdr Labelled .bil}

\item{num_bands}{Number of bandwidths}

\item{n_times}{Number of models}

\item{BI_part}{}

\item{BI_tot}{}

\item{SD_BI_part}{}

\item{SD_BI_tot}{}
}
\value{
\code{selfinfo_mod_}, \code{info_mod_} and \code{info_mod_means_} (all followed by the name of the species). The first two tables are merely informative about how the modelling process has been developed and the results of each model. Whereas \code{info_mod_means_} shows the means of the n models run for each bandwidth
}
\description{
It aims at (1) defining what is the minimum or optimal background extent necessary to fit good partial SDMs and/or (2) determining if the background area used to fit a partial SDM is reliable enough to extract ecologically relevant conclusions from it.
}
\details{
minba aims at (1) defining what is the minimum or optimal background extent 
necessary to fit good partial SDMs and/or (2) determining if the background 
area used to fit a partial SDM is reliable enough to extract ecologically 
relevant conclusions from it.
}
\author{
Xavier Rotllan-Puig
}
